refactor(codex-rules): split parser helpers
